requirements.txt
alembic==1.18.4
argon2-cffi==25.1.0
asyncpg==0.30.0
casbin_sqlalchemy_adapter==1.4.0
cryptography==46.0.3
email-validator==2.3.0
fastapi==0.119.0
passlib==1.7.4
psycopg2==2.9.11
pydantic_core==2.41.4
pydantic-settings==2.12.0
pydantic==2.12.2
python-jose==3.5.0
rich==14.2.0
secure==1.0.1
starlette==0.48.0
uvicorn==0.38.0
